requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.11.8
    # via requests
chardet==3.0.4
    # via requests
idna==2.10
    # via requests
requests==2.25.0
    # via -r requirements.in
ruamel.yaml==0.16.12
    # via -r requirements.in
urllib3==1.26.5
    # via requests
